New User-Article relation

+++ app/controllers/articles_controller.rb
@@ -8,11 +8,14 @@ class ArticlesController < ApplicationController
   end
 
   def new
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
+
     @article = Article.new
   end
 
   def create
     @article = Article.new(article_params)
+    @article.user = current_user
 
     if @article.save
       redirect_to @article
@@ -22,7 +25,10 @@ class ArticlesController < ApplicationController
   end
 
   def edit
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
+
     @article = Article.find(params[:id])
+    session_notice(:danger, 'Wrong User') unless equal_with_current_user?(@article.user)
   end
 
   def update
@@ -36,10 +42,16 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    article = Article.find(params[:id])
-    article.destroy
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
 
-    redirect_to articles_path
+    article = Article.find(params[:id])
+
+    if equal_with_current_user?(article.user)
+      article.destroy
+      redirect_to articles_path
+    else
+      session_notice(:danger, 'Wrong User')
+    end
   end
 
   private
